refactor(frontend): move GldtStaking component
