+++ src/frontend/src/icp/components/staking/gldt/GldtStaking.svelte
